external payment generate

+++ api/src/main/groovy/bff/model/MoneyInput.groovy
@@ -5,5 +5,5 @@ import groovy.transform.Immutable
 @Immutable
 class MoneyInput {
     String currency
-    BigDecimal amount
+    Double amount
 }
